Slider is there, but does not rerender states
