+++ frontend/package.json
@@ -14,6 +14,7 @@
     "@deck.gl/react": "^9.1.11",
     "d3-color": "^3.1.0",
     "d3-scale": "^4.0.2",
+    "date-fns": "^4.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-map-gl": "^8.0.4"
